extracted interfaces to it's own file
