+++ src/components/editor/editor-window/interfaces.ts
@@ -1,0 +1,15 @@
+import CodeMirror from 'codemirror'
+
+export interface SelectionData {
+  ranges: AnchorAndHead[]
+}
+
+interface AnchorAndHead {
+  anchor: CodeMirror.Position
+  head: CodeMirror.Position
+}
+
+export interface Positions {
+  startPosition: CodeMirror.Position
+  endPosition: CodeMirror.Position
+}
